Added Tower, Pathing, Map Layout
In this I've managed to make a basic map layout to begin working on other mechanics eg the Pathing and tower placements. So far the Pathing is pretty much completed just need to do some more testing on it down the line after I finish up with the enemy. Towers are in progress. They currently follow a enemy when they come into range and track it until its out of range. Next plans are to add in towers shooting and a possible update with the map visuals and sprites of enemies.