Adding Currency, New Tower & Menu
Added in the currency from killing enemies and then spending it to place more towers. This will be changed from how much currency is dropped and how much the towers will cost. Also put in a menu for the user to buy towers. This is just temporary as there were different plans for the menu buying being a popup. Added in a catapult tower which needs to be fixed up as projectiles still uses arrows rather than a ball.