UI Changes
I've changed the tower buy menu and added images for each tower. It's still a work in progress as i'm attempting it to be a pop up when the user clicks on the build tower tile. Also changed the Enemies layer as there was a problem of them clipping behind some tiles at times.