Finished Pause Button
I've finished off the pause menu allowing the user to pause the game freezing the game and able to resume it when the play button is clicked. The Exit doesn't do anything as of right now as I'm going to implement that when my team mate has completed his task of making the level screen(which consists of a overall map of missions).