Level 3 Update
Almost finished with layout of design of level 3. Also found a problem when the user has the buy menu open for towers and then clicking on a different plot without closing the menu causes both plots involved to be unusable for the rest of the game. So in need of fixing that asap.