Upgraded Tower Targeting
I updated/upgraded the towers targeting ensuring it always targets the enemy in the front and preventing other enemies from body blocking projectiles from reaching the target. I touched up the mage towers design fixing the crystal from rotating as it looked weird before. For that I've changed the Base image to include the crystal as a still image and hidden the crystal itself where the targeting point is set.