AI Wave System Start
Begun working on the AI integration with the wave system. Increasing and decreasing the difficulty depending on the players progression. So far it has a basic format that works where it goes off how much HP the user has. I also noticed a few mistakes in level 1 and 2 where enemy 4 wasn't in the spawn list so added those in and change in level 1 the cost of a catapult which was incorrect at first.