Working on the AI
So far its a work in progress, Im having issues with it counting enemies failed to go halfway making the next wave more difficult even when the enemies have reached the end or all gone past half way which should keep the next wave as base wave or decrease the wave spawns to help the player. So just working on fixing that.