Basic Star System
When the user wins the game when the win screen pops up it displays how many stars they have earned. How many stars they get depends on their hp the more hp you have the more stars you have. I will be implementing this into the map system so the user can then use these stars to buy perma upgrades for the towers but for now I'm still waiting on my team member to complete the scene for it.