AI Changes and Fixing Bugs
I've been trying to fix the AI as its inconsistent sometimes and still working on getting that fixed. I've fixed a problem with pause menus where the user while on the pause screen could still click on the speed up button and resume the game while the pause menu is going. So fixed that to prevent the player from doing that. Also added in a FPS counter in the top left so I could see if the games performance is running well and because some users prefer to have a fps counter while playing.