AI Update and Balancing
Huge update to the AI, changing how it works. I've improved on the wave balancing to make it more accurate. The AI now adjusts the variant enemies spawn rates. If the player is doing well it will bring in more enemy variants from more tanky and speedy making it harder or if the player is struggling it spawns less of them and more of the basic weaker ones. Also did some balancing from Towers stats and enemy stats, decreasing the coin drops changing a few of the hps. And for towers changed the dps and updated the upgrade system so when a tower is upgraded it also upgrades the total damage it now does. Also increased the total upgrade cost.